Bug 476025 - Powerpc, update expected results for the vbit-test
In commit:

  commit 26a3abd27db2ef63dafea1ecab00e8239f341f0f
  Author: Eyal Soha <eyalsoha@gmail.com>
  Date:   Thu Feb 10 09:52:54 2022 -0700

      Bug 432801 - Valgrind 3.16.1 reports a jump based on uninitialized
      memory somehow related to clang and signals

      Add support for precise computation of SIMD greater-than on
      amd64 and x86.

      This adds support for 64bit, 16bit, and 8bit to the existing 32bit
      support.

The Iop_CmpGT64Ux2 is only supported on PPC32 and PPC64.  The above
commit adds a more precise method, expensiveCmpGT, for setting the
vibits for the Iop_CmpGT64Ux2 Iop.

The expected results for the vbit test were not updated to the new more
precise results.  This patch updates the expected results for the Iop.

Patch tested on Power 10.