ltp tests: Use new LTP 20250530, allow for filtering valgrind output
Use fresh LTP 20250530.

Allow for filtering valgrind output:

In some cases the LTP tests intentionally work with SIGSEGV.  This
happens e.g. with the mmap18 and select03 testcases.  Valgrind
detects SIGSEGV and reports that as a failure.

Such report can't be suppressed using the suppressions mechanism.
That's why this update comes with "output filters". Filters are
scripts that read from their stdin, and write filtered output to
their stdout. Filters reside in auxprogs/filters.

This update comes with 2 filters: For mmap18, and for select03.
They are awk scripts.

Except for filters, this update also blacklists testcase fork13
because it is slow.  It is possible to add comments prefixed with
the '#' sign (implicitly - because they don't match any testcase
name), so a comment is added too.

This update also introduces new default valgrind --vgdb=no switch. It
improves valgrind behavior with nftw01 nftw6401 setfsgid04 setfsgid03_16
and symlink03 testcases. These were previously complaining like this:

==22969== could not unlink /tmp/vgdb-pipe-from-vgdb-to-22969-by-root-on ...