Wrap linux specific syscall sysfs
The sysfs syscall is deprecated, but in some cases it may still
be used.  The Linux Test Project covers it.

The (obsolete) sysfs() system call returns information about the
filesystem types currently present in the kernel.  The specific
form of the sysfs() call and the information returned depends on
the option in effect:

1  Translate the filesystem identifier string fsname into a
   filesystem type index.

2  Translate the filesystem type index fs_index into a null-
   terminated filesystem identifier string.  This string will be
   written to the buffer pointed to by buf.  Make sure that buf
   has enough space to accept the string.

3  Return the total number of filesystem types currently present
   in the kernel.

Declare a sys_sysfs wrapper in priv_syswrap-linux.h and hook it
for {amd64,arm,mips32,mips64,ppc32,ppc64,s390x,x86}-linux
using LINXY with PRE and POST handler in syswrap-linux.c

https://bugs.kde.org/show_bug.cgi?id=506929